Fixed  expmidd types
