--- types/expmidd/Middleware/cors.d.ts
@@ -1,7 +1,0 @@
-import { ICors } from '../typescript/cors.definitions';
-declare class AddCors {
-    constructor();
-    cors: ICors;
-}
-declare const _default: AddCors;
-export default _default;
