--- types/expmidd/Middleware/express.d.ts
@@ -1,3 +1,0 @@
-import { Request, Response } from 'express';
-declare const runMiddleware: (req: Request, res: Response, fn: any) => Promise<unknown>;
-export default runMiddleware;
