--- types/expmidd/constants/cors.d.ts
@@ -1,2 +1,0 @@
-import { IHTTP_METHODS } from '../typescript/cors.definitions';
-export declare const HTTP_METHODS: IHTTP_METHODS;
