+++ types/expmidd/package.json
@@ -2,7 +2,8 @@
     "private": true,
     "dependencies": {
         "cors": "^2.8.5",
-        "expmidd": "0.3.0"
+        "expmidd": "0.2.0",
+        "express": "^4.17.1"
     },
     "types": "index",
     "typesVersions": {
@@ -18,6 +19,7 @@
         }
     },
     "devDependencies": {
-        "@types/cors": "^2.8.10"
+        "@types/cors": "^2.8.10",
+        "@types/express": "^4.17.12"
     }
 }
