+++ types/expmidd/tslint.json
@@ -1,3 +1,1 @@
-{
-    "extends": "dtslint/dt.json"
-}
+{ "extends": "dtslint/dt.json" }
